backend/requirements.txt
# Backend requirements (API server)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart
python-dotenv

# LLM & Training
torch>=2.1.0
transformers>=4.38.0
datasets>=2.14,<3
peft>=0.7.0
trl>=0.7.10
accelerate>=0.27.0
bitsandbytes>=0.41.0

# RAG
sentence-transformers>=2.5.0
chromadb>=0.4.22

# LangChain (multi-agent framework)
langchain>=0.1.0
langchain-community>=0.0.10

# Tools
numexpr
tavily-python
requests
tqdm
